Renamed config package and removed HK2 bindings
