+++ hermes-frontend/src/main/java/pl/allegro/tech/hermes/frontend/config/CommonConfiguration.java
@@ -1,4 +1,4 @@
-package pl.allegro.tech.hermes.frontend.di.config;
+package pl.allegro.tech.hermes.frontend.config;
 
 import com.codahale.metrics.MetricRegistry;
 import com.fasterxml.jackson.databind.ObjectMapper;
