+++ hermes-frontend/src/main/java/pl/allegro/tech/hermes/frontend/config/TopicMetadataLoadingStartupHookCondition.java
@@ -1,4 +1,4 @@
-package pl.allegro.tech.hermes.frontend.di.config;
+package pl.allegro.tech.hermes.frontend.config;
 
 import org.springframework.context.annotation.Condition;
 import org.springframework.context.annotation.ConditionContext;
@@ -7,13 +7,13 @@ import pl.allegro.tech.hermes.common.config.ConfigFactory;
 
 import java.util.Objects;
 
-import static pl.allegro.tech.hermes.common.config.Configs.FRONTEND_STARTUP_TOPIC_SCHEMA_LOADING_ENABLED;
+import static pl.allegro.tech.hermes.common.config.Configs.FRONTEND_STARTUP_TOPIC_METADATA_LOADING_ENABLED;
 
-public class TopicSchemaLoadingStartupHookCondition implements Condition {
+public class TopicMetadataLoadingStartupHookCondition implements Condition {
 
     @Override
     public boolean matches(ConditionContext context, AnnotatedTypeMetadata metadata) {
         ConfigFactory config = Objects.requireNonNull(context.getBeanFactory()).getBean(ConfigFactory.class);
-        return config.getBooleanProperty(FRONTEND_STARTUP_TOPIC_SCHEMA_LOADING_ENABLED);
+        return config.getBooleanProperty(FRONTEND_STARTUP_TOPIC_METADATA_LOADING_ENABLED);
     }
 }
